fix(sync): float->short UB sweep — same class as the COB desync bug
Audit sweep of out-of-range float->short conversions in synced code
(IPathController stopTurnHeading / turn-rate clamps / return value,
HoverAirMoveType + GroundMoveType direct-control heading,
LuaSyncedMoveCtrl.SetHeading, LuaSyncedRead.GetFacingFromHeading).
All now truncate via int32: defined behavior, identical to what the
fleet's gcc-x86 builds compute.

+++ rts/Lua/LuaSyncedMoveCtrl.cpp
@@ -463,7 +463,7 @@ int LuaSyncedMoveCtrl::SetHeading(lua_State* L)
 	if (moveType == nullptr)
 		return 0;
 
-	const short heading = (short)luaL_checknumber(L, 2);
+	const short heading = (short)(int)luaL_checknumber(L, 2);
 	ASSERT_SYNCED((short)heading);
 	moveType->SetHeading(heading);
 	return 0;
